PersonalRagServer/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-pinecone
pinecone
huggingface_hub
python-dotenv
requests
httpx
google-generativeai
langchain-google-genai